refactor macro utils, skip field support
